Move InterleavedContent to api/common/content_types.py
