+++ llama_stack/apis/common/deployment_types.py
@@ -12,16 +12,6 @@ from llama_models.schema_utils import json_schema_type
 from pydantic import BaseModel
 
 
-@json_schema_type(
-    schema={"type": "string", "format": "uri", "pattern": "^(https?://|file://|data:)"}
-)
-class URL(BaseModel):
-    uri: str
-
-    def __str__(self) -> str:
-        return self.uri
-
-
 @json_schema_type
 class RestAPIMethod(Enum):
     GET = "GET"
