+++ llama_stack/providers/utils/datasetio/url_utils.py
@@ -10,7 +10,7 @@ from urllib.parse import unquote
 
 import pandas
 
-from llama_stack.apis.common.deployment_types import URL
+from llama_stack.apis.common.content_types import URL
 
 from llama_stack.providers.utils.memory.vector_store import parse_data_url
