+++ llama_stack/providers/utils/memory/vector_store.py
@@ -21,7 +21,7 @@ from pypdf import PdfReader
 from llama_models.llama3.api.datatypes import *  # noqa: F403
 from llama_models.llama3.api.tokenizer import Tokenizer
 
-from llama_stack.apis.inference import InterleavedContent, TextContentItem
+from llama_stack.apis.common.content_types import InterleavedContent, TextContentItem
 from llama_stack.apis.memory import *  # noqa: F403
 from llama_stack.apis.memory_banks import VectorMemoryBank
 from llama_stack.providers.datatypes import Api
